Reorganized server code into files.
