+++ docs/server/rest/databases/test.db
@@ -1,16 +1,8 @@
-{"type":"aircraft","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","_company":"tGwfy0QbiHL6lRJN","notes":"new aircraft","_id":"3cySRjFd4AUGY3g5","createdAt":{"$$date":1598819110593},"updatedAt":{"$$date":1598819110607}}
-{"type":"contact","name":"","address":"","city":"","state":"","zip":"","phone":"","email":"","notes":"new contact","_id":"7izwBKCRI4R4TCNj","createdAt":{"$$date":1598819110503},"updatedAt":{"$$date":1598819110503}}
-{"type":"aircraft","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","_company":"xe142BW40oITxGxu","notes":"new aircraft","_id":"Ps8111J6BDi5s4pS","createdAt":{"$$date":1598822130989},"updatedAt":{"$$date":1598822131002}}
-{"type":"contact","name":"","address":"","city":"","state":"","zip":"","phone":"","email":"","notes":"new contact","_id":"SiUmtWxmZHMCZxja","createdAt":{"$$date":1598822130920},"updatedAt":{"$$date":1598822130920}}
-{"type":"engine","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","time_since_overhaul":"","_aircraft":"Ps8111J6BDi5s4pS","notes":"new engine","_id":"aHaQn5nXeGSZXWsM","createdAt":{"$$date":1598822130982},"updatedAt":{"$$date":1598822131026}}
-{"type":"engine","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","time_since_overhaul":"","_aircraft":"3cySRjFd4AUGY3g5","notes":"new engine","_id":"piQQ6ipDIlO5pMWD","createdAt":{"$$date":1598819110585},"updatedAt":{"$$date":1598819110639}}
-{"type":"company","name":"kim","address":"","city":"","state":"","zip":"","phone":"","email":"","_contacts":["7izwBKCRI4R4TCNj"],"notes":"","_id":"tGwfy0QbiHL6lRJN","createdAt":{"$$date":1598819110461},"updatedAt":{"$$date":1598819110612}}
-{"type":"company","name":"kim","address":"","city":"","state":"","zip":"","phone":"","email":"","_contacts":["SiUmtWxmZHMCZxja"],"notes":"","_id":"xe142BW40oITxGxu","createdAt":{"$$date":1598822130897},"updatedAt":{"$$date":1598822131007}}
-{"type":"company","name":"kim","address":"","city":"","state":"","zip":"","phone":"","email":"","_contacts":[],"notes":"","_id":"K4QmMSvqV3QGIdJi","createdAt":{"$$date":1598823157351},"updatedAt":{"$$date":1598823157351}}
-{"type":"contact","name":"","address":"","city":"","state":"","zip":"","phone":"","email":"","notes":"new contact","_id":"Gv6lvbgcmK5oDj0i","createdAt":{"$$date":1598823157368},"updatedAt":{"$$date":1598823157368}}
-{"type":"company","name":"kim","address":"","city":"","state":"","zip":"","phone":"","email":"","_contacts":["Gv6lvbgcmK5oDj0i"],"notes":"","_id":"K4QmMSvqV3QGIdJi","createdAt":{"$$date":1598823157351},"updatedAt":{"$$date":1598823157375}}
-{"type":"engine","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","time_since_overhaul":"","_aircraft":"","notes":"new engine","_id":"FEMO4udh9PYg6272","createdAt":{"$$date":1598823157424},"updatedAt":{"$$date":1598823157424}}
-{"type":"aircraft","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","_company":"","notes":"new aircraft","_id":"5NeGIr2HsCSgs3EC","createdAt":{"$$date":1598823157435},"updatedAt":{"$$date":1598823157435}}
-{"type":"aircraft","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","_company":"K4QmMSvqV3QGIdJi","notes":"new aircraft","_id":"5NeGIr2HsCSgs3EC","createdAt":{"$$date":1598823157435},"updatedAt":{"$$date":1598823157439}}
-{"type":"company","name":"kim","address":"","city":"","state":"","zip":"","phone":"","email":"","_contacts":["Gv6lvbgcmK5oDj0i"],"notes":"","_id":"K4QmMSvqV3QGIdJi","createdAt":{"$$date":1598823157351},"updatedAt":{"$$date":1598823157443}}
-{"type":"engine","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","time_since_overhaul":"","_aircraft":"5NeGIr2HsCSgs3EC","notes":"new engine","_id":"FEMO4udh9PYg6272","createdAt":{"$$date":1598823157424},"updatedAt":{"$$date":1598823157456}}
+{"type":"company","name":"kim","address":"","city":"","state":"","zip":"","phone":"","email":"","_contacts":[],"notes":"","_id":"Dvcl4iSPKUSSqSW6","createdAt":{"$$date":1598826833252},"updatedAt":{"$$date":1598826833252}}
+{"type":"contact","name":"","address":"","city":"","state":"","zip":"","phone":"","email":"","notes":"new cl.contact","_id":"rv9FjqGkj4195Juk","createdAt":{"$$date":1598826833270},"updatedAt":{"$$date":1598826833270}}
+{"type":"company","name":"kim","address":"","city":"","state":"","zip":"","phone":"","email":"","_contacts":["rv9FjqGkj4195Juk"],"notes":"","_id":"Dvcl4iSPKUSSqSW6","createdAt":{"$$date":1598826833252},"updatedAt":{"$$date":1598826833278}}
+{"type":"engine","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","time_since_overhaul":"","_aircraft":"","notes":"new cl.engine","_id":"ASah6tINyWMtDnUK","createdAt":{"$$date":1598826833332},"updatedAt":{"$$date":1598826833332}}
+{"type":"aircraft","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","_company":"","notes":"new cl.aircraft","_id":"ZptPJh6TAVumUuyX","createdAt":{"$$date":1598826833339},"updatedAt":{"$$date":1598826833339}}
+{"type":"aircraft","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","_company":"Dvcl4iSPKUSSqSW6","notes":"new cl.aircraft","_id":"ZptPJh6TAVumUuyX","createdAt":{"$$date":1598826833339},"updatedAt":{"$$date":1598826833344}}
+{"type":"company","name":"kim","address":"","city":"","state":"","zip":"","phone":"","email":"","_contacts":["rv9FjqGkj4195Juk"],"notes":"","_id":"Dvcl4iSPKUSSqSW6","createdAt":{"$$date":1598826833252},"updatedAt":{"$$date":1598826833347}}
+{"type":"engine","name":"","model":"","make":"","serial_no":"","registration_no":"","time_in_service":"","time_since_overhaul":"","_aircraft":"ZptPJh6TAVumUuyX","notes":"new cl.engine","_id":"ASah6tINyWMtDnUK","createdAt":{"$$date":1598826833332},"updatedAt":{"$$date":1598826833360}}
